Section_098_-_Day_98_-_Professional_Portfolio_Project_-_[Python_Automation]/4_job_automation_web/requirements.txt
Flask
pandas
selenium
pyyaml
python-dotenv